#7309: move a missing-attribute test into string.eo to make room
